# Implement an `AB_TEST()` macro for automated performance comparisons between alternative versions of a code block