@Zireael07 2023-02-10
I recall that same user doing some stuff threaded depending on something, and wondering about performance?